+++ app/src/main/java/com/kotlinisgood/boomerang/ui/audiorecord/Audio.kt
@@ -1,4 +1,4 @@
-package com.kotlinisgood.boomerang.ui.voicerecord
+package com.kotlinisgood.boomerang.ui.audiorecord
 
 data class Audio(val path: String, val duration: Int, val recognizedText: String)
