+++ app/src/main/java/com/kotlinisgood/boomerang/ui/audiorecord/AudioRecordViewModel.kt
@@ -1,4 +1,4 @@
-package com.kotlinisgood.boomerang.ui.voicerecord
+package com.kotlinisgood.boomerang.ui.audiorecord
 
 import androidx.lifecycle.ViewModel
