+++ app/src/main/java/com/kotlinisgood/boomerang/ui/audiorecord/VoiceRecordFragment.kt
@@ -1,4 +1,4 @@
-package com.kotlinisgood.boomerang.ui.voicerecord
+package com.kotlinisgood.boomerang.ui.audiorecord
 
 import android.Manifest
 import android.content.Intent
